+++ smac/main/smbo.py
@@ -14,7 +14,9 @@ from smac.acquisition.function.abstract_acquisition_function import (
     AbstractAcquisitionFunction,
 )
 from smac.callback.callback import Callback
+from smac.constants import ASK_AND_TELL_VALID_WARN_MODES
 from smac.intensifier.abstract_intensifier import AbstractIntensifier
+from smac.main.exceptions import AskAndTellBudgetExhaustedError
 from smac.model.abstract_model import AbstractModel
 from smac.runhistory import StatusType, TrialInfo, TrialValue
 from smac.runhistory.runhistory import RunHistory
@@ -67,6 +69,7 @@ class SMBO:
         runhistory: RunHistory,
         intensifier: AbstractIntensifier,
         overwrite: bool = False,
+        warn_mode: str = "warn_always",
     ):
         self._scenario = scenario
         self._configspace = scenario.configspace
@@ -76,15 +79,23 @@ class SMBO:
         self._runner = runner
         self._overwrite = overwrite
 
+        if warn_mode not in ASK_AND_TELL_VALID_WARN_MODES:
+            raise ValueError(f"Unknown warn_mode `{warn_mode}`. Allowed: {sorted(ASK_AND_TELL_VALID_WARN_MODES)}")
+
+        self._warn_mode = warn_mode
+
         # Internal variables
         self._finished = False
         self._stop = False  # Gracefully stop SMAC
         self._callbacks: list[Callback] = []
+        self._warned_on_ask_after_budget_exhausted = False
 
         # Stats variables
-        self._start_time: float | None = None
         self._used_target_function_walltime = 0.0
         self._used_target_function_cputime = 0.0
+
+        # Start the timer. In case of resuming an optimization process, the starting time is set by the load method
+        self._start_time: float = time.time()
 
         # Set walltime used method for intensifier
         self._intensifier.used_walltime = lambda: self.used_walltime  # type: ignore
@@ -155,6 +166,26 @@ class SMBO:
             Information about the trial (config, instance, seed, budget).
         """
         logger.debug("Calling ask...")
+
+        if self.budget_exhausted:
+            message = (
+                "ask() was called after the scenario budget was exhausted."
+                f" (remaining wallclock time: {self.remaining_walltime}, "
+                f"remaining cpu time: {self.remaining_cputime}, "
+                f"remaining trials: {self.remaining_trials}). "
+                "SMAC will continue returning trials for backward compatibility."
+            )
+
+            if self._warn_mode == "exception":
+                raise AskAndTellBudgetExhaustedError(message)
+            elif self._warn_mode == "warn_never":
+                pass
+            elif self._warn_mode == "warn_once":
+                if not self._warned_on_ask_after_budget_exhausted:
+                    logger.warning(message)
+                    self._warned_on_ask_after_budget_exhausted = True
+            elif self._warn_mode == "warn_always":
+                logger.warning(message)
 
         for callback in self._callbacks:
             callback.on_ask_start(self)
@@ -284,11 +315,6 @@ class SMBO:
             else:
                 return self.intensifier.get_incumbents()
 
-        # Start the timer before we do anything
-        # If we continue the optimization, the starting time is set by the load method
-        if self._start_time is None:
-            self._start_time = time.time()
-
         for callback in self._callbacks:
             callback.on_start(self)
 
@@ -372,6 +398,8 @@ class SMBO:
         self._used_target_function_walltime = 0
         self._used_target_function_cputime = 0
         self._finished = False
+        self._warned_on_ask_after_budget_exhausted = False
+        self._start_time = time.time()
 
         # We also reset runhistory and intensifier here
         self._runhistory.reset()